% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{geom_stripes}
\alias{geom_stripes}
\title{Alternating Background Colour}
\usage{
geom_stripes(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Add alternating background color along the y-axis. The geom takes default
aesthetics \code{odd} and \code{even} that receive color codes. The codes
would preferably be in the 8-hex ARGB format to allow for transparency if
the geom is meant to be used as visual background.
}
