% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_markdowns.R
\name{render_markdowns}
\alias{render_markdowns}
\title{Render all helper modals, introduction, and notes documents}
\usage{
render_markdowns(.output_dir = "inst/app/www")
}
\arguments{
\item{.output_dir}{character, directory location for output markdowns}
}
\value{
nothing
}
\description{
Render all helper modals, introduction, and notes documents
}
\examples{
\dontrun{
render_markdowns()
}
}
